% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{accu.consts}
\alias{accu.consts}
\title{Define constangs values}
\format{An object of class \code{list} of length 29.}
\usage{
accu.consts
}
\description{
Define constangs values
}
\keyword{datasets}
