% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-client.R
\name{accu_client}
\alias{accu_client}
\title{Initialize an Accu modeler Client}
\usage{
accu_client(tracking_uri = NULL)
}
\arguments{
\item{tracking_uri}{The tracking URI. If not provided, defaults to the service
set by \code{accu_set_tracking_uri()}.}
}
\description{
Initializes and returns an Accu modeler client.
}
