% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-experiments.R
\name{accu_create_experiment}
\alias{accu_create_experiment}
\title{Create Experiment}
\usage{
accu_create_experiment(artifact_location = NULL, client = NULL)
}
\arguments{
\item{artifact_location}{Location where all artifacts for this experiment are stored. If
not provided, the remote server will select an appropriate default.}

\item{name}{The name of the experiment to create.}
}
\description{
Creates an modeler experiment and returns its id.
}
