% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-globals.R
\name{accu_set_tracking_uri}
\alias{accu_set_tracking_uri}
\title{Set Remote Tracking URI}
\usage{
accu_set_tracking_uri()
}
\arguments{
\item{uri}{The URI to the remote server.}
}
\description{
Specifies the URI to the remote server that will be used
to create experiments.
}
