% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{add_experiment}
\alias{add_experiment}
\title{add_experiment Function}
\usage{
add_experiment(model_name, traindata = NULL, testdata)
}
\arguments{
\item{model_name:}{your model name}

\item{testdata:}{test data for computing roc, pr}
}
\description{
Add you experiment to list of model experiments when using accuinsight modeler service.
}
