% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/os_getenv.R
\name{get_os_env}
\alias{get_os_env}
\title{Read .env file from user home directory}
\usage{
get_os_env(env_type = "LC", env_path = NULL, env_file = NULL)
}
\description{
Read .env file from user home directory
}
